requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
yt-dlp>=2024.1.0
openai-whisper>=20231117
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# ─── Audio ────────────────────────────────────────────────────────────────────
pydub>=0.25.1
sounddevice>=0.4.6

# ─── TTS ──────────────────────────────────────────────────────────────────────
edge-tts>=7.0.0

# ─── LLM ──────────────────────────────────────────────────────────────────────
groq>=0.4.0

# ─── Backend ──────────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ─── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.30.0
requests>=2.31.0
python-multipart>=0.0.9
